Updated sample data return based on API Updates
We've changed the precip service a bunch of times, so this reflects the new output for the given request. Precipitation requests now return precip amounts based on hourly METAR observations, instead of 6-hourly. There are other flags that can be added to the request which will change even more how the accumulations are derived. 

The fundamental effect of this update changes the resulting total precip over the given period from .13" to .09"